refactor: migrate pentest and operator-dashboard to shared hooks
Replace inline message accumulation in both components with the
useAgentMessages hook. Extract usePentestOrchestration and useAssetPolling
hooks from pentest.tsx. Clean up operator-dashboard dead state, fix
unsafe AbortError detection, normalize abort controller to useRef.

pentest.tsx: 1374 → 872 lines (37% reduction)
operator-dashboard: 511 → 440 lines (14% reduction)

Refs: #199
