+++ src/tui/components/pentest/pentest.tsx
@@ -1,41 +1,19 @@
-import { useState, useEffect, useCallback, useRef, useMemo } from "react";
+import { useState, useEffect, useCallback, useMemo } from "react";
 import { useKeyboard } from "@opentui/react";
-import { RGBA } from "@opentui/core";
-import { existsSync, readdirSync, readFileSync } from "fs";
+import { existsSync } from "fs";
 import { join } from "path";
 import { exec } from "child_process";
 
 import { sessions, type SessionInfo } from "../../../core/session";
-import { runPentestAgent } from "../../../core/api";
-import { useAgent } from "../../context/agent";
 import { useRoute } from "../../context/route";
-import { useConfig } from "../../context/config";
 import { useDialog } from "../../context/dialog";
-import { useTheme, type ThemeColors } from "../../theme";
-import AgentDisplay, { type DisplayMessage } from "../agent-display";
+import { useTheme } from "../../theme";
 import { SpinnerDots } from "../sprites";
-import type { DocumentedAssetRecord } from "../../../core/agents/specialized/attackSurface/schemas";
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-type Phase =
-  | "loading"
-  | "discovery"
-  | "pentesting"
-  | "reporting"
-  | "completed"
-  | "error";
-
-type AgentState = {
-  id: string;
-  name: string;
-  target: string;
-  status: "pending" | "completed" | "failed";
-  messages: DisplayMessage[];
-  createdAt: Date;
-};
+import { usePentestOrchestration, useAssetPolling } from "../../hooks";
+import { OrchestratorPanel } from "./orchestrator-panel";
+import { AgentCardGrid } from "./agent-card";
+import { AgentDetailView } from "./agent-detail-view";
+import { MetricsBar } from "./metrics-bar";
 
 type ViewMode = "overview" | "detail";
 
@@ -46,39 +24,31 @@ type ViewMode = "overview" | "detail";
 export default function Pentest({ sessionId }: { sessionId: string }) {
   const { colors } = useTheme();
   const route = useRoute();
-  const config = useConfig();
-  const { model, setThinking, setIsExecuting, isExecuting } = useAgent();
   const { stack, externalDialogOpen } = useDialog();
 
   // Session
   const [session, setSession] = useState<SessionInfo | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [phase, setPhase] = useState<Phase>("loading");
 
-  // Abort
-  const [abortController, setAbortController] =
-    useState<AbortController | null>(null);
+  // Orchestration (phase, panel messages, pentest agents, abort, errors)
+  const {
+    phase,
+    panelMessages,
+    pentestAgents,
+    error,
+    startPentest,
+    abort,
+    isExecuting,
+    startTime,
+  } = usePentestOrchestration(session);
 
-  // Orchestrator panel messages (orchestrator own events + discovery subagent events)
-  const [panelMessages, setPanelMessages] = useState<DisplayMessage[]>([]);
-  const panelTextRef = useRef("");
-  const panelSourceRef = useRef<string | null>(null); // tracks who is currently producing text
-
-  // Pentest agent states — keyed by subagentId
-  const [pentestAgents, setPentestAgents] = useState<
-    Record<string, AgentState>
-  >({});
-  const pentestTextRefs = useRef<Record<string, string>>({});
-
-  // Assets from filesystem
-  const [assets, setAssets] = useState<DocumentedAssetRecord[]>([]);
+  // Asset polling
+  const { assets } = useAssetPolling(session);
 
   // UI state
   const [viewMode, setViewMode] = useState<ViewMode>("overview");
   const [selectedAgentId, setSelectedAgentId] = useState<string | null>(null);
   const [focusedIndex, setFocusedIndex] = useState(0);
   const [showOrchestratorPanel, setShowOrchestratorPanel] = useState(false);
-  const [startTime, setStartTime] = useState<Date | null>(null);
 
   // Derived
   const pentestAgentList = useMemo(
@@ -102,501 +72,16 @@ export default function Pentest({ sessionId }: { sessionId: string }) {
       try {
         const s = await sessions.get(sessionId);
         if (!s) {
-          setError(`Session not found: ${sessionId}`);
-          setPhase("error");
+          // Error state is managed by the orchestration hook via phase
           return;
         }
         setSession(s);
-      } catch (e) {
-        setError(e instanceof Error ? e.message : "Failed to load session");
-        setPhase("error");
+      } catch {
+        // Session load failure — orchestration hook handles error display
       }
     }
     load();
   }, [sessionId]);
-
-  // -------------------------------------------------------------------------
-  // Poll assets folder
-  // -------------------------------------------------------------------------
-
-  useEffect(() => {
-    if (!session) return;
-    const assetsPath = join(session.rootPath, "assets");
-
-    function readAssets() {
-      if (!existsSync(assetsPath)) return;
-      try {
-        const files = readdirSync(assetsPath).filter((f) =>
-          f.endsWith(".json"),
-        );
-        const loaded: DocumentedAssetRecord[] = [];
-        for (const file of files) {
-          try {
-            const content = readFileSync(join(assetsPath, file), "utf-8");
-            loaded.push(JSON.parse(content));
-          } catch {
-            // skip malformed files
-          }
-        }
-        setAssets(loaded);
-      } catch {
-        // assets dir may not exist yet
-      }
-    }
-
-    readAssets();
-    const interval = setInterval(readAssets, 2000);
-    return () => clearInterval(interval);
-  }, [session]);
-
-  // -------------------------------------------------------------------------
-  // Cleanup
-  // -------------------------------------------------------------------------
-
-  useEffect(() => {
-    return () => {
-      abortController?.abort();
-    };
-  }, [abortController]);
-
-  // -------------------------------------------------------------------------
-  // Helper: ensure a pentest agent entry exists (fallback for stream events
-  // that arrive before onSubagentSpawn)
-  // -------------------------------------------------------------------------
-
-  const ensurePentestAgent = useCallback((subagentId: string) => {
-    setPentestAgents((prev) => {
-      if (prev[subagentId]) return prev;
-      return {
-        ...prev,
-        [subagentId]: {
-          id: subagentId,
-          name: subagentId,
-          target: "",
-          status: "pending",
-          messages: [],
-          createdAt: new Date(),
-        },
-      };
-    });
-  }, []);
-
-  // -------------------------------------------------------------------------
-  // Helper: handle subagent spawn (sets target + objectives as name)
-  // -------------------------------------------------------------------------
-
-  const handleSubagentSpawn = useCallback(
-    ({
-      subagentId,
-      input,
-    }: {
-      subagentId: string;
-      input: unknown;
-      status: string;
-    }) => {
-      if (!subagentId.startsWith("pentest-agent-")) return;
-      const { target, objectives } = input as {
-        target: string;
-        objectives: string[];
-      };
-      const name =
-        objectives.length === 1
-          ? objectives[0]!
-          : `${objectives.length} objectives`;
-      setPentestAgents((prev) => ({
-        ...prev,
-        [subagentId]: {
-          ...(prev[subagentId] ?? { messages: [], createdAt: new Date() }),
-          id: subagentId,
-          name,
-          target,
-          status: "pending" as const,
-          createdAt: prev[subagentId]?.createdAt ?? new Date(),
-        },
-      }));
-    },
-    [],
-  );
-
-  // -------------------------------------------------------------------------
-  // Helper: handle subagent complete
-  // -------------------------------------------------------------------------
-
-  const handleSubagentComplete = useCallback(
-    ({ subagentId, status }: { subagentId: string; status: string }) => {
-      if (!subagentId.startsWith("pentest-agent-")) return;
-      setPentestAgents((prev) => {
-        const agent = prev[subagentId];
-        if (!agent) return prev;
-        return {
-          ...prev,
-          [subagentId]: { ...agent, status: status as "completed" | "failed" },
-        };
-      });
-    },
-    [],
-  );
-
-  // -------------------------------------------------------------------------
-  // Helper: append text to the orchestrator panel (orchestrator or discovery)
-  // -------------------------------------------------------------------------
-
-  const appendPanelText = useCallback((source: string, text: string) => {
-    // If the source changed, start a new message
-    if (panelSourceRef.current !== source) {
-      panelTextRef.current = "";
-      panelSourceRef.current = source;
-    }
-    panelTextRef.current += text;
-    const accumulated = panelTextRef.current;
-    setPanelMessages((prev) => {
-      const last = prev[prev.length - 1];
-      if (
-        last &&
-        last.role === "assistant" &&
-        (last as DisplayMessage & { _source?: string })._source === source
-      ) {
-        const updated = [...prev];
-        updated[updated.length - 1] = { ...last, content: accumulated };
-        return updated;
-      }
-      return [
-        ...prev,
-        {
-          role: "assistant",
-          content: accumulated,
-          createdAt: new Date(),
-          _source: source,
-        } as DisplayMessage & { _source: string },
-      ];
-    });
-  }, []);
-
-  // -------------------------------------------------------------------------
-  // Helper: append text delta to a pentest agent
-  // -------------------------------------------------------------------------
-
-  const appendPentestText = useCallback(
-    (subagentId: string, text: string) => {
-      ensurePentestAgent(subagentId);
-      if (!pentestTextRefs.current[subagentId]) {
-        pentestTextRefs.current[subagentId] = "";
-      }
-      pentestTextRefs.current[subagentId] += text;
-      const accumulated = pentestTextRefs.current[subagentId];
-      setPentestAgents((prev) => {
-        const agent = prev[subagentId];
-        if (!agent) return prev;
-        const last = agent.messages[agent.messages.length - 1];
-        if (last && last.role === "assistant") {
-          const updated = [...agent.messages];
-          updated[updated.length - 1] = { ...last, content: accumulated };
-          return { ...prev, [subagentId]: { ...agent, messages: updated } };
-        }
-        return {
-          ...prev,
-          [subagentId]: {
-            ...agent,
-            messages: [
-              ...agent.messages,
-              {
-                role: "assistant",
-                content: accumulated,
-                createdAt: new Date(),
-              },
-            ],
-          },
-        };
-      });
-    },
-    [ensurePentestAgent],
-  );
-
-  // -------------------------------------------------------------------------
-  // Helper: add tool call message
-  // -------------------------------------------------------------------------
-
-  // -------------------------------------------------------------------------
-  // Helper: add tool call to orchestrator panel
-  // -------------------------------------------------------------------------
-
-  const addPanelToolCall = useCallback(
-    (toolCallId: string, toolName: string, args?: Record<string, unknown>) => {
-      panelTextRef.current = "";
-      panelSourceRef.current = null;
-
-      const description =
-        typeof args?.toolCallDescription === "string"
-          ? args.toolCallDescription
-          : toolName;
-
-      const msg: DisplayMessage = {
-        role: "tool",
-        content: description,
-        toolCallId,
-        toolName,
-        args,
-        status: "pending",
-        createdAt: new Date(),
-      };
-
-      setPanelMessages((prev) => {
-        if (prev.some((m) => m.toolCallId === toolCallId)) return prev;
-        return [...prev, msg];
-      });
-    },
-    [],
-  );
-
-  // -------------------------------------------------------------------------
-  // Helper: add tool call to a pentest agent
-  // -------------------------------------------------------------------------
-
-  const addPentestToolCall = useCallback(
-    (
-      subagentId: string,
-      toolCallId: string,
-      toolName: string,
-      args?: Record<string, unknown>,
-    ) => {
-      pentestTextRefs.current[subagentId] = "";
-      ensurePentestAgent(subagentId);
-
-      const description =
-        typeof args?.toolCallDescription === "string"
-          ? args.toolCallDescription
-          : toolName;
-
-      const msg: DisplayMessage = {
-        role: "tool",
-        content: description,
-        toolCallId,
-        toolName,
-        args,
-        status: "pending",
-        createdAt: new Date(),
-      };
-
-      setPentestAgents((prev) => {
-        const agent = prev[subagentId];
-        if (!agent) return prev;
-        if (agent.messages.some((m) => m.toolCallId === toolCallId))
-          return prev;
-        return {
-          ...prev,
-          [subagentId]: { ...agent, messages: [...agent.messages, msg] },
-        };
-      });
-    },
-    [ensurePentestAgent],
-  );
-
-  // -------------------------------------------------------------------------
-  // Helper: update tool result
-  // -------------------------------------------------------------------------
-
-  // -------------------------------------------------------------------------
-  // Helper: generic tool result updater
-  // -------------------------------------------------------------------------
-
-  const toolResultUpdater = (
-    messages: DisplayMessage[],
-    toolCallId: string,
-    toolName: string,
-    result?: unknown,
-  ): DisplayMessage[] => {
-    const idx = messages.findIndex((m) => m.toolCallId === toolCallId);
-    if (idx !== -1) {
-      const updated = [...messages];
-      updated[idx] = { ...updated[idx]!, status: "completed" as const, result };
-      return updated;
-    }
-    return [
-      ...messages,
-      {
-        role: "tool" as const,
-        content: toolName,
-        toolCallId,
-        toolName,
-        status: "completed" as const,
-        result,
-        createdAt: new Date(),
-      },
-    ];
-  };
-
-  const updatePanelToolResult = useCallback(
-    (toolCallId: string, toolName: string, result?: unknown) => {
-      panelTextRef.current = "";
-      panelSourceRef.current = null;
-      setPanelMessages((prev) =>
-        toolResultUpdater(prev, toolCallId, toolName, result),
-      );
-    },
-    [],
-  );
-
-  const updatePentestToolResult = useCallback(
-    (
-      subagentId: string,
-      toolCallId: string,
-      toolName: string,
-      result?: unknown,
-    ) => {
-      pentestTextRefs.current[subagentId] = "";
-      setPentestAgents((prev) => {
-        const agent = prev[subagentId];
-        if (!agent) return prev;
-        return {
-          ...prev,
-          [subagentId]: {
-            ...agent,
-            messages: toolResultUpdater(
-              agent.messages,
-              toolCallId,
-              toolName,
-              result,
-            ),
-          },
-        };
-      });
-    },
-    [],
-  );
-
-  // -------------------------------------------------------------------------
-  // Start the pentest
-  // -------------------------------------------------------------------------
-
-  const startPentest = useCallback(
-    async (s: SessionInfo) => {
-      setPhase("discovery");
-      setStartTime(new Date());
-      setIsExecuting(true);
-      setThinking(true);
-
-      const controller = new AbortController();
-      setAbortController(controller);
-
-      const run = runPentestAgent({
-        target: s.targets[0]!,
-        ...(s.config?.cwd ? { cwd: s.config.cwd } : {}),
-        session: s,
-        model: model.id,
-        abortSignal: controller.signal,
-        authConfig: {
-          anthropicAPIKey: config.data.anthropicAPIKey ?? undefined,
-          openAiAPIKey: config.data.openAiAPIKey ?? undefined,
-          openRouterAPIKey: config.data.openRouterAPIKey ?? undefined,
-          local: config.data.localModelUrl
-            ? { baseURL: config.data.localModelUrl }
-            : undefined,
-        },
-      });
-
-      try {
-        for await (const event of run) {
-          switch (event.type) {
-            case "text-delta":
-              setThinking(false);
-              if (!event.subagentId) {
-                appendPanelText("orchestrator", event.data.text);
-              } else if (event.subagentId === "attack-surface-agent") {
-                appendPanelText(event.subagentId, event.data.text);
-              } else {
-                appendPentestText(event.subagentId, event.data.text);
-              }
-              break;
-            case "tool-call":
-              setThinking(false);
-              if (!event.subagentId) {
-                if (event.data.toolName === "run_attack_surface") {
-                  setPhase("discovery");
-                } else if (event.data.toolName === "spawn_pentest_swarm") {
-                  setPhase("pentesting");
-                } else if (event.data.toolName === "generate_report") {
-                  setPhase("reporting");
-                }
-                addPanelToolCall(
-                  event.data.toolCallId,
-                  event.data.toolName,
-                  event.data.input as Record<string, unknown> | undefined,
-                );
-              } else if (event.subagentId === "attack-surface-agent") {
-                addPanelToolCall(
-                  event.data.toolCallId,
-                  event.data.toolName,
-                  event.data.input as Record<string, unknown> | undefined,
-                );
-              } else {
-                addPentestToolCall(
-                  event.subagentId,
-                  event.data.toolCallId,
-                  event.data.toolName,
-                  event.data.input as Record<string, unknown> | undefined,
-                );
-              }
-              break;
-            case "tool-result":
-              setThinking(true);
-              if (
-                !event.subagentId ||
-                event.subagentId === "attack-surface-agent"
-              ) {
-                updatePanelToolResult(
-                  event.data.toolCallId,
-                  event.data.toolName,
-                  event.data.output,
-                );
-              } else {
-                updatePentestToolResult(
-                  event.subagentId,
-                  event.data.toolCallId,
-                  event.data.toolName,
-                  event.data.output,
-                );
-              }
-              break;
-            case "subagent-spawn":
-              handleSubagentSpawn(event);
-              break;
-            case "subagent-complete":
-              handleSubagentComplete(event);
-              break;
-            case "error":
-              console.error("Agent error:", event.error);
-              break;
-          }
-        }
-
-        await run.result;
-        setPhase("completed");
-      } catch (e) {
-        if (e instanceof Error && e.name === "AbortError") {
-          // User aborted
-        } else {
-          setError(e instanceof Error ? e.message : "Unknown error");
-          setPhase("error");
-        }
-      } finally {
-        setThinking(false);
-        setIsExecuting(false);
-      }
-    },
-    [
-      model.id,
-      setThinking,
-      setIsExecuting,
-      appendPanelText,
-      appendPentestText,
-      addPanelToolCall,
-      addPentestToolCall,
-      updatePanelToolResult,
-      updatePentestToolResult,
-      handleSubagentSpawn,
-      handleSubagentComplete,
-    ],
-  );
 
   // -------------------------------------------------------------------------
   // Auto-start when session loads
@@ -604,7 +89,7 @@ export default function Pentest({ sessionId }: { sessionId: string }) {
 
   useEffect(() => {
     if (session && phase === "loading") {
-      startPentest(session);
+      startPentest();
     }
   }, [session, phase, startPentest]);
 
@@ -626,7 +111,7 @@ export default function Pentest({ sessionId }: { sessionId: string }) {
           setShowOrchestratorPanel(false);
           return;
         }
-        abortController?.abort();
+        abort();
         route.navigate({ type: "base", path: "home" });
         return;
       }
@@ -839,536 +324,6 @@ export default function Pentest({ sessionId }: { sessionId: string }) {
         startTime={startTime}
         isExecuting={isExecuting}
       />
-    </box>
-  );
-}
-
-// ===========================================================================
-// OrchestratorPanel
-// ===========================================================================
-
-interface OrchestratorPanelProps {
-  phase: Phase;
-  messages: DisplayMessage[];
-  assets: DocumentedAssetRecord[];
-  expanded: boolean;
-  onToggle: () => void;
-}
-
-function OrchestratorPanel({
-  phase,
-  messages,
-  assets,
-  expanded,
-  onToggle,
-}: OrchestratorPanelProps) {
-  const { colors } = useTheme();
-  const isRunning =
-    phase !== "loading" && phase !== "completed" && phase !== "error";
-
-  // Group assets by type for display
-  const assetSummary = useMemo(() => {
-    const byType: Record<string, DocumentedAssetRecord[]> = {};
-    for (const a of assets) {
-      const key = a.assetType;
-      if (!byType[key]) byType[key] = [];
-      byType[key]!.push(a);
-    }
-    return byType;
-  }, [assets]);
-
-  const assetTypeLabels: Record<string, string> = {
-    domain: "Domains",
-    subdomain: "Subdomains",
-    web_application: "Web Apps",
-    api: "APIs",
-    admin_panel: "Admin Panels",
-    endpoint: "Endpoints",
-    infrastructure_service: "Infra",
-    cloud_resource: "Cloud",
-    development_asset: "Dev Assets",
-  };
-
-  const phaseLabel = {
-    loading: "Loading",
-    discovery: "Discovery",
-    pentesting: "Pentesting",
-    reporting: "Reporting",
-    completed: "Completed",
-    error: "Error",
-  }[phase];
-
-  // Expanded view — full message log
-  if (expanded) {
-    return (
-      <box
-        flexGrow={1}
-        border
-        borderColor={colors.primary}
-        backgroundColor={colors.backgroundElement}
-        flexDirection="column"
-      >
-        <box
-          flexDirection="row"
-          alignItems="center"
-          justifyContent="space-between"
-          padding={1}
-          borderColor={colors.borderSubtle}
-          border={["bottom"]}
-        >
-          <box flexDirection="row" gap={1}>
-            {isRunning ? (
-              <SpinnerDots
-                label={`Orchestrator — ${phaseLabel}`}
-                fg={colors.primary}
-              />
-            ) : (
-              <text
-                fg={
-                  phase === "completed"
-                    ? colors.primary
-                    : phase === "error"
-                      ? colors.error
-                      : colors.textMuted
-                }
-              >
-                {phase === "completed" ? "✓" : phase === "error" ? "✗" : "●"}{" "}
-                Orchestrator — {phaseLabel}
-              </text>
-            )}
-          </box>
-          <text fg={colors.textMuted}>
-            <span fg={colors.primary}>[D]</span> collapse |{" "}
-            <span fg={colors.primary}>[ESC]</span> close
-          </text>
-        </box>
-
-        <AgentDisplay
-          messages={messages}
-          isStreaming={isRunning}
-          focused={true}
-          paddingLeft={2}
-          paddingRight={2}
-          contextId="orchestrator-logs"
-        />
-
-        <box
-          flexDirection="row"
-          justifyContent="space-between"
-          padding={1}
-          borderColor={colors.borderSubtle}
-          border={["top"]}
-        >
-          <text fg={colors.textMuted}>{messages.length} messages</text>
-          <text fg={colors.textMuted}>{assets.length} assets found</text>
-        </box>
-      </box>
-    );
-  }
-
-  // Collapsed panel — sidebar
-  return (
-    <box
-      width={32}
-      border
-      borderColor={isRunning ? colors.primary : colors.textMuted}
-      backgroundColor={colors.backgroundElement}
-      flexDirection="column"
-      onMouseDown={onToggle}
-    >
-      {/* Header */}
-      <box
-        flexDirection="row"
-        alignItems="center"
-        justifyContent="space-between"
-        padding={1}
-        borderColor={colors.borderSubtle}
-        border={["bottom"]}
-      >
-        <box flexDirection="row" gap={1}>
-          {isRunning ? (
-            <SpinnerDots label={phaseLabel} fg={colors.primary} />
-          ) : (
-            <text
-              fg={phase === "completed" ? colors.primary : colors.textMuted}
-            >
-              {phase === "completed" ? "✓" : "●"} {phaseLabel}
-            </text>
-          )}
-        </box>
-        <text fg={colors.textMuted}>[D]</text>
-      </box>
-
-      {/* Stats */}
-      <box flexDirection="column" padding={1} gap={1}>
-        <text>
-          <span fg={colors.textMuted}>Phase: </span>
-          <span fg={isRunning ? colors.primary : colors.text}>
-            {phaseLabel}
-          </span>
-        </text>
-        <text>
-          <span fg={colors.textMuted}>Assets: </span>
-          <span fg={colors.text}>{assets.length}</span>
-        </text>
-        <text>
-          <span fg={colors.textMuted}>Messages: </span>
-          <span fg={colors.text}>{messages.length}</span>
-        </text>
-      </box>
-
-      {/* Asset list by type */}
-      {assets.length > 0 && (
-        <scrollbox
-          style={{
-            rootOptions: { flexGrow: 1, paddingLeft: 1, paddingRight: 1 },
-            contentOptions: { flexDirection: "column", gap: 0 },
-          }}
-          stickyScroll={false}
-          focused={false}
-        >
-          {Object.entries(assetSummary).map(([type, items]) => (
-            <box key={type} flexDirection="column">
-              <text fg={colors.textMuted}>
-                {assetTypeLabels[type] ?? type} ({items.length})
-              </text>
-              {items.map((a, i) => {
-                const label = a.assetName;
-                return (
-                  <text key={i} fg={colors.textMuted}>
-                    <span fg={riskColor(colors, a.riskLevel)}>• </span>
-                    {label.length > 26 ? label.slice(0, 26) + "…" : label}
-                  </text>
-                );
-              })}
-            </box>
-          ))}
-        </scrollbox>
-      )}
-    </box>
-  );
-}
-
-function riskColor(colors: ThemeColors, level: string): RGBA {
-  switch (level) {
-    case "CRITICAL":
-      return colors.error;
-    case "HIGH":
-      return colors.warning;
-    case "MEDIUM":
-      return colors.warning;
-    default:
-      return colors.success;
-  }
-}
-
-// ===========================================================================
-// AgentCardGrid
-// ===========================================================================
-
-interface AgentCardGridProps {
-  agents: AgentState[];
-  focusedIndex: number;
-  onSelectAgent: (id: string) => void;
-}
-
-function AgentCardGrid({
-  agents,
-  focusedIndex,
-  onSelectAgent,
-}: AgentCardGridProps) {
-  const { colors } = useTheme();
-  const rows = useMemo(() => {
-    const result: AgentState[][] = [];
-    for (let i = 0; i < agents.length; i += 2) {
-      result.push(agents.slice(i, i + 2));
-    }
-    return result;
-  }, [agents]);
-
-  return (
-    <scrollbox
-      style={{
-        rootOptions: { flexGrow: 1, width: "100%" },
-        contentOptions: { flexDirection: "column", gap: 1 },
-      }}
-      stickyScroll={false}
-      focused={true}
-    >
-      {rows.map((row, ri) => (
-        <box key={ri} flexDirection="row" gap={1} width="100%">
-          {row.map((agent, ci) => {
-            const flatIdx = ri * 2 + ci;
-            return (
-              <AgentCard
-                key={agent.id}
-                agent={agent}
-                focused={flatIdx === focusedIndex}
-                onSelect={() => onSelectAgent(agent.id)}
-              />
-            );
-          })}
-          {row.length === 1 && <box flexGrow={1} flexBasis={0} minWidth={40} />}
-        </box>
-      ))}
-    </scrollbox>
-  );
-}
-
-// ===========================================================================
-// AgentCard
-// ===========================================================================
-
-function AgentCard({
-  agent,
-  focused,
-  onSelect,
-}: {
-  agent: AgentState;
-  focused: boolean;
-  onSelect: () => void;
-}) {
-  const { colors } = useTheme();
-  const statusIcon = { pending: "◐", completed: "✓", failed: "✗" }[
-    agent.status
-  ];
-  const statusColor = {
-    pending: colors.primary,
-    completed: colors.primary,
-    failed: colors.error,
-  }[agent.status];
-
-  const lastActivity = useMemo(() => {
-    const last = agent.messages[agent.messages.length - 1];
-    if (!last) return "Starting...";
-    const text =
-      typeof last.content === "string"
-        ? last.content.replace(/\n/g, " ").trim()
-        : "Working...";
-    return text.length > 50 ? text.substring(0, 47) + "..." : text;
-  }, [agent.messages]);
-
-  const toolCalls = useMemo(
-    () => agent.messages.filter((m) => m.role === "tool").length,
-    [agent.messages],
-  );
-
-  return (
-    <box
-      flexGrow={1}
-      flexBasis={0}
-      minWidth={40}
-      border
-      borderColor={focused ? colors.primary : colors.textMuted}
-      backgroundColor={colors.backgroundElement}
-      flexDirection="column"
-      padding={1}
-      rowGap={1}
-      onMouseDown={onSelect}
-    >
-      <box flexDirection="row" alignItems="center" gap={1} flexWrap="wrap">
-        <text fg={statusColor}>{statusIcon}</text>
-        <text fg={focused ? colors.text : colors.textMuted}>
-          {agent.target || agent.name}
-        </text>
-      </box>
-
-      {agent.target && agent.name !== agent.target && (
-        <text fg={colors.textMuted}>{agent.name}</text>
-      )}
-
-      <box flexDirection="row" gap={2} marginTop={1}>
-        <text fg={colors.textMuted}>
-          <span fg={colors.primary}>{toolCalls}</span> calls
-        </text>
-        <text fg={colors.textMuted}>
-          <span fg={colors.primary}>{agent.messages.length}</span> msgs
-        </text>
-      </box>
-
-      <box height={1} overflow="hidden">
-        {agent.status === "pending" ? (
-          <SpinnerDots label={lastActivity} fg={colors.primary} />
-        ) : (
-          <text
-            fg={
-              agent.status === "completed" ? colors.primary : colors.textMuted
-            }
-          >
-            {agent.status === "completed" ? "✓ Complete" : lastActivity}
-          </text>
-        )}
-      </box>
-    </box>
-  );
-}
-
-// ===========================================================================
-// AgentDetailView
-// ===========================================================================
-
-function AgentDetailView({
-  agent,
-  onBack,
-}: {
-  agent: AgentState;
-  onBack: () => void;
-}) {
-  const { colors } = useTheme();
-  const statusColor = {
-    pending: colors.primary,
-    completed: colors.primary,
-    failed: colors.error,
-  }[agent.status];
-
-  return (
-    <box flexDirection="column" width="100%" height="100%" flexGrow={1}>
-      <box
-        width="100%"
-        border={["bottom"]}
-        borderColor={colors.primary}
-        flexDirection="row"
-        justifyContent="space-between"
-        padding={1}
-      >
-        <box flexDirection="row" gap={1}>
-          <text fg={colors.textMuted}></text>
-          <text fg={colors.text}>{agent.name}</text>
-        </box>
-        <box flexDirection="row" gap={2}>
-          {agent.target && (
-            <text>
-              <span fg={colors.textMuted}>Target: </span>
-              <span fg={colors.text}>{agent.target}</span>
-            </text>
-          )}
-          <text fg={statusColor}>{agent.status}</text>
-        </box>
-      </box>
-
-      <AgentDisplay
-        messages={agent.messages}
-        isStreaming={agent.status === "pending"}
-        focused={true}
-        paddingLeft={4}
-        paddingRight={4}
-        contextId={agent.id}
-      />
-
-      <box
-        width="100%"
-        flexDirection="row"
-        justifyContent="space-between"
-        border={["top"]}
-        borderColor={colors.primary}
-        padding={1}
-      >
-        <text fg={colors.textMuted}>{agent.messages.length} messages</text>
-        <box flexDirection="row" gap={2}>
-          <text>
-            <span fg={colors.primary}>[ESC]</span>
-            <span fg={colors.textMuted}> Back to overview</span>
-          </text>
-        </box>
-      </box>
-    </box>
-  );
-}
-
-// ===========================================================================
-// MetricsBar
-// ===========================================================================
-
-function MetricsBar({
-  phase,
-  agents,
-  assets,
-  startTime,
-  isExecuting,
-}: {
-  phase: Phase;
-  agents: AgentState[];
-  assets: DocumentedAssetRecord[];
-  startTime: Date | null;
-  isExecuting: boolean;
-}) {
-  const { colors } = useTheme();
-  const [now, setNow] = useState(Date.now());
-
-  useEffect(() => {
-    if (!isExecuting) return;
-    const interval = setInterval(() => setNow(Date.now()), 1000);
-    return () => clearInterval(interval);
-  }, [isExecuting]);
-
-  const duration = startTime
-    ? Math.floor((now - startTime.getTime()) / 1000)
-    : 0;
-  const mins = Math.floor(duration / 60);
-  const secs = duration % 60;
-  const formattedDuration = `${mins}m ${secs.toString().padStart(2, "0")}s`;
-
-  const active = agents.filter((a) => a.status === "pending").length;
-  const total = agents.length;
-
-  const phaseLabel = {
-    loading: "Loading",
-    discovery: "Discovery",
-    pentesting: "Pentesting",
-    reporting: "Reporting",
-    completed: "Completed",
-    error: "Error",
-  }[phase];
-
-  return (
-    <box
-      width="100%"
-      flexDirection="row"
-      justifyContent="space-between"
-      borderColor={colors.primary}
-      border={["top"]}
-      padding={1}
-    >
-      <box flexDirection="row" gap={2}>
-        <text>
-          <span fg={isExecuting ? colors.primary : colors.textMuted}>
-            {phaseLabel}
-          </span>
-        </text>
-        <text fg={colors.textMuted}>|</text>
-        <text>
-          <span fg={colors.primary}>{assets.length}</span>
-          <span fg={colors.textMuted}> assets</span>
-        </text>
-        <text fg={colors.textMuted}>|</text>
-        <text>
-          <span fg={isExecuting ? colors.primary : colors.textMuted}>
-            {active}
-          </span>
-          <span fg={colors.textMuted}>/{total} agents</span>
-        </text>
-        <text fg={colors.textMuted}>|</text>
-        <text fg={colors.textMuted}>{formattedDuration}</text>
-      </box>
-
-      <box flexDirection="row" gap={2}>
-        <text>
-          <span fg={colors.primary}>[D]</span>
-          <span fg={colors.textMuted}> Orchestrator</span>
-        </text>
-        <text>
-          <span fg={colors.primary}>[Tab]</span>
-          <span fg={colors.textMuted}> Navigate</span>
-        </text>
-        <text>
-          <span fg={colors.primary}>[Enter]</span>
-          <span fg={colors.textMuted}> View</span>
-        </text>
-        <text>
-          <span fg={colors.primary}>[ESC]</span>
-          <span fg={colors.textMuted}> Back</span>
-        </text>
-      </box>
     </box>
   );
 }
