+++ src/tui/hooks/use-pentest-orchestration.ts
@@ -1,0 +1,434 @@
+/**
+ * usePentestOrchestration Hook
+ *
+ * Encapsulates the pentest event stream consumption, phase state machine,
+ * panel message accumulation (via useAgentMessages), pentest subagent
+ * management, and abort handling. Extracted from pentest.tsx to reduce
+ * the component's hook count and separate orchestration logic from rendering.
+ */
+
+import { useState, useCallback, useRef, useEffect } from "react";
+import { runPentestAgent } from "../../core/api";
+import { useAgent } from "../context/agent";
+import { useConfig } from "../context/config";
+import { buildAuthConfig } from "../utils";
+import { useAgentMessages, toolResultUpdater } from "./use-agent-messages";
+import type { SessionInfo } from "../../core/session";
+import type { DisplayMessage } from "../components/agent-display";
+
+// ---------------------------------------------------------------------------
+// Types
+// ---------------------------------------------------------------------------
+
+export type Phase =
+  | "loading"
+  | "discovery"
+  | "pentesting"
+  | "reporting"
+  | "completed"
+  | "error";
+
+export type AgentState = {
+  id: string;
+  name: string;
+  target: string;
+  status: "pending" | "completed" | "failed";
+  messages: DisplayMessage[];
+  createdAt: Date;
+};
+
+export interface UsePentestOrchestrationReturn {
+  phase: Phase;
+  panelMessages: DisplayMessage[];
+  pentestAgents: Record<string, AgentState>;
+  error: string | null;
+  startPentest: () => Promise<void>;
+  abort: () => void;
+  isExecuting: boolean;
+  startTime: Date | null;
+}
+
+// ---------------------------------------------------------------------------
+// Hook
+// ---------------------------------------------------------------------------
+
+export function usePentestOrchestration(
+  session: SessionInfo | null,
+): UsePentestOrchestrationReturn {
+  const { model, setThinking, setIsExecuting, isExecuting } = useAgent();
+  const config = useConfig();
+
+  // Phase state machine
+  const [phase, setPhase] = useState<Phase>("loading");
+  const [error, setError] = useState<string | null>(null);
+  const [startTime, setStartTime] = useState<Date | null>(null);
+
+  // Abort — useRef instead of useState to avoid stale closure issues
+  const abortControllerRef = useRef<AbortController | null>(null);
+
+  // Orchestrator panel messages (orchestrator + discovery subagent events)
+  const {
+    messages: panelMessages,
+    appendText: appendPanelTextRaw,
+    addToolCall: addPanelToolCall,
+    updateToolResult: updatePanelToolResult,
+  } = useAgentMessages();
+
+  // Pentest agent states — keyed by subagentId
+  const [pentestAgents, setPentestAgents] = useState<
+    Record<string, AgentState>
+  >({});
+  const pentestTextRefs = useRef<Record<string, string>>({});
+
+  // -------------------------------------------------------------------------
+  // Abort cleanup on unmount
+  // -------------------------------------------------------------------------
+
+  useEffect(() => {
+    return () => {
+      abortControllerRef.current?.abort();
+    };
+  }, []);
+
+  // -------------------------------------------------------------------------
+  // Abort function exposed to consumers
+  // -------------------------------------------------------------------------
+
+  const abort = useCallback(() => {
+    abortControllerRef.current?.abort();
+  }, []);
+
+  // -------------------------------------------------------------------------
+  // Helper: ensure a pentest agent entry exists
+  // -------------------------------------------------------------------------
+
+  const ensurePentestAgent = useCallback((subagentId: string) => {
+    setPentestAgents((prev) => {
+      if (prev[subagentId]) return prev;
+      return {
+        ...prev,
+        [subagentId]: {
+          id: subagentId,
+          name: subagentId,
+          target: "",
+          status: "pending",
+          messages: [],
+          createdAt: new Date(),
+        },
+      };
+    });
+  }, []);
+
+  // -------------------------------------------------------------------------
+  // Helper: handle subagent spawn
+  // -------------------------------------------------------------------------
+
+  const handleSubagentSpawn = useCallback(
+    ({
+      subagentId,
+      input,
+    }: {
+      subagentId: string;
+      input: unknown;
+      status: string;
+    }) => {
+      if (!subagentId.startsWith("pentest-agent-")) return;
+      const { target, objectives } = input as {
+        target: string;
+        objectives: string[];
+      };
+      const name =
+        objectives.length === 1
+          ? objectives[0]!
+          : `${objectives.length} objectives`;
+      setPentestAgents((prev) => ({
+        ...prev,
+        [subagentId]: {
+          ...(prev[subagentId] ?? { messages: [], createdAt: new Date() }),
+          id: subagentId,
+          name,
+          target,
+          status: "pending" as const,
+          createdAt: prev[subagentId]?.createdAt ?? new Date(),
+        },
+      }));
+    },
+    [],
+  );
+
+  // -------------------------------------------------------------------------
+  // Helper: handle subagent complete
+  // -------------------------------------------------------------------------
+
+  const handleSubagentComplete = useCallback(
+    ({ subagentId, status }: { subagentId: string; status: string }) => {
+      if (!subagentId.startsWith("pentest-agent-")) return;
+      setPentestAgents((prev) => {
+        const agent = prev[subagentId];
+        if (!agent) return prev;
+        return {
+          ...prev,
+          [subagentId]: {
+            ...agent,
+            status: status as "completed" | "failed",
+          },
+        };
+      });
+    },
+    [],
+  );
+
+  // -------------------------------------------------------------------------
+  // Pentest agent text helper
+  // -------------------------------------------------------------------------
+
+  const appendPentestText = useCallback(
+    (subagentId: string, text: string) => {
+      ensurePentestAgent(subagentId);
+      if (!pentestTextRefs.current[subagentId]) {
+        pentestTextRefs.current[subagentId] = "";
+      }
+      pentestTextRefs.current[subagentId] += text;
+      const accumulated = pentestTextRefs.current[subagentId];
+      setPentestAgents((prev) => {
+        const agent = prev[subagentId];
+        if (!agent) return prev;
+        const last = agent.messages[agent.messages.length - 1];
+        if (last && last.role === "assistant") {
+          const updated = [...agent.messages];
+          updated[updated.length - 1] = { ...last, content: accumulated };
+          return { ...prev, [subagentId]: { ...agent, messages: updated } };
+        }
+        return {
+          ...prev,
+          [subagentId]: {
+            ...agent,
+            messages: [
+              ...agent.messages,
+              {
+                role: "assistant",
+                content: accumulated,
+                createdAt: new Date(),
+              },
+            ],
+          },
+        };
+      });
+    },
+    [ensurePentestAgent],
+  );
+
+  // -------------------------------------------------------------------------
+  // Pentest agent tool call helper
+  // -------------------------------------------------------------------------
+
+  const addPentestToolCall = useCallback(
+    (
+      subagentId: string,
+      toolCallId: string,
+      toolName: string,
+      args?: Record<string, unknown>,
+    ) => {
+      pentestTextRefs.current[subagentId] = "";
+      ensurePentestAgent(subagentId);
+
+      const description =
+        typeof args?.toolCallDescription === "string"
+          ? args.toolCallDescription
+          : toolName;
+
+      const msg: DisplayMessage = {
+        role: "tool",
+        content: description,
+        toolCallId,
+        toolName,
+        args,
+        status: "pending",
+        createdAt: new Date(),
+      };
+
+      setPentestAgents((prev) => {
+        const agent = prev[subagentId];
+        if (!agent) return prev;
+        if (agent.messages.some((m) => m.toolCallId === toolCallId))
+          return prev;
+        return {
+          ...prev,
+          [subagentId]: { ...agent, messages: [...agent.messages, msg] },
+        };
+      });
+    },
+    [ensurePentestAgent],
+  );
+
+  // -------------------------------------------------------------------------
+  // Pentest agent tool result helper
+  // -------------------------------------------------------------------------
+
+  const updatePentestToolResult = useCallback(
+    (
+      subagentId: string,
+      toolCallId: string,
+      toolName: string,
+      result?: unknown,
+    ) => {
+      pentestTextRefs.current[subagentId] = "";
+      setPentestAgents((prev) => {
+        const agent = prev[subagentId];
+        if (!agent) return prev;
+        return {
+          ...prev,
+          [subagentId]: {
+            ...agent,
+            messages: toolResultUpdater(
+              agent.messages,
+              toolCallId,
+              toolName,
+              result,
+            ),
+          },
+        };
+      });
+    },
+    [],
+  );
+
+  // -------------------------------------------------------------------------
+  // Start the pentest — the main event loop
+  // -------------------------------------------------------------------------
+
+  const startPentest = useCallback(async () => {
+    if (!session) return;
+
+    setPhase("discovery");
+    setStartTime(new Date());
+    setIsExecuting(true);
+    setThinking(true);
+
+    const controller = new AbortController();
+    abortControllerRef.current = controller;
+
+    const run = runPentestAgent({
+      target: session.targets[0]!,
+      ...(session.config?.cwd ? { cwd: session.config.cwd } : {}),
+      session,
+      model: model.id,
+      abortSignal: controller.signal,
+      authConfig: buildAuthConfig(config.data),
+    });
+
+    try {
+      for await (const event of run) {
+        switch (event.type) {
+          case "text-delta":
+            setThinking(false);
+            if (!event.subagentId) {
+              appendPanelTextRaw(event.data.text);
+            } else if (event.subagentId === "attack-surface-agent") {
+              appendPanelTextRaw(event.data.text);
+            } else {
+              appendPentestText(event.subagentId, event.data.text);
+            }
+            break;
+          case "tool-call":
+            setThinking(false);
+            if (!event.subagentId) {
+              if (event.data.toolName === "run_attack_surface") {
+                setPhase("discovery");
+              } else if (event.data.toolName === "spawn_pentest_swarm") {
+                setPhase("pentesting");
+              } else if (event.data.toolName === "generate_report") {
+                setPhase("reporting");
+              }
+              addPanelToolCall(
+                event.data.toolCallId,
+                event.data.toolName,
+                event.data.input as Record<string, unknown> | undefined,
+              );
+            } else if (event.subagentId === "attack-surface-agent") {
+              addPanelToolCall(
+                event.data.toolCallId,
+                event.data.toolName,
+                event.data.input as Record<string, unknown> | undefined,
+              );
+            } else {
+              addPentestToolCall(
+                event.subagentId,
+                event.data.toolCallId,
+                event.data.toolName,
+                event.data.input as Record<string, unknown> | undefined,
+              );
+            }
+            break;
+          case "tool-result":
+            setThinking(true);
+            if (
+              !event.subagentId ||
+              event.subagentId === "attack-surface-agent"
+            ) {
+              updatePanelToolResult(
+                event.data.toolCallId,
+                event.data.toolName,
+                event.data.output,
+              );
+            } else {
+              updatePentestToolResult(
+                event.subagentId,
+                event.data.toolCallId,
+                event.data.toolName,
+                event.data.output,
+              );
+            }
+            break;
+          case "subagent-spawn":
+            handleSubagentSpawn(event);
+            break;
+          case "subagent-complete":
+            handleSubagentComplete(event);
+            break;
+          case "error":
+            console.error("Agent error:", event.error);
+            break;
+        }
+      }
+
+      await run.result;
+      setPhase("completed");
+    } catch (e) {
+      if (e instanceof Error && e.name === "AbortError") {
+        // User aborted — not an error
+      } else {
+        setError(e instanceof Error ? e.message : "Unknown error");
+        setPhase("error");
+      }
+    } finally {
+      setThinking(false);
+      setIsExecuting(false);
+    }
+  }, [
+    session,
+    model.id,
+    config.data,
+    setThinking,
+    setIsExecuting,
+    appendPanelTextRaw,
+    appendPentestText,
+    addPanelToolCall,
+    addPentestToolCall,
+    updatePanelToolResult,
+    updatePentestToolResult,
+    handleSubagentSpawn,
+    handleSubagentComplete,
+  ]);
+
+  return {
+    phase,
+    panelMessages,
+    pentestAgents,
+    error,
+    startPentest,
+    abort,
+    isExecuting,
+    startTime,
+  };
+}
